sandboxify: Reverse child/parent logic to preserve exec chain.
Avoid a double fork by making the sandboxify process exec into the
target binary and have the child instead act as the tracer.

This ensures that the return status of the executed binary is the the
return of the whole command line.
It also allows proper supervision without hack (ala systemd
`Type=forking`) as the launcher of the command line can simply
waitpid().

The reversal of the logic works around the kernel's yama protection by
having the parent first grant the child PTRACE access. This gives time
to the child to setup the trace and suspend SECCOMP.
Both threads are coordinated through a pipe which serves to pause the
parent while the child sets up PTRACE and SECCOMP but also allows to
detect death of the child and properly abort.